CLN move benchmark feature tests+FIX test run
