+++ benchopt/tests/test_benchmarks.py
@@ -1,12 +1,9 @@
 import pytest
 import numpy as np
 
-import benchopt
-from benchopt.cli.main import run
 from benchopt.runner import _Callback
 from benchopt.stopping_criterion import SAMPLING_STRATEGIES
 from benchopt.utils import product_param
-from benchopt.utils.temp_benchmark import temp_benchmark
 
 
 def test_benchmark_objective(benchmark, dataset_simu):
@@ -222,68 +219,3 @@ def _test_solver_one_objective(solver, objective):
 
             diff = val_eps - val_star
             assert diff >= 0
-
-
-def test_deprecated_stopping_strategy():
-    # XXX remove in 1.5
-    assert benchopt.__version__ < '1.5'
-
-    solver1 = """from benchopt import BaseSolver
-    import numpy as np
-
-    class Solver(BaseSolver):
-        name = 'solver1'
-        stopping_strategy = 'iteration'
-
-        def run(self, n_iter): pass
-
-        def set_objective(self, X, y, lmbd):
-            self.n_features = X.shape[1]
-
-        def get_result(self, **data):
-            return {'beta': np.zeros(self.n_features)}
-    """
-
-    solver2 = solver1.replace("stopping_strategy", "sampling_strategy")
-    solver2 = solver2.replace("solver1", "solver2")
-
-    with temp_benchmark(solvers=[solver1, solver2]) as benchmark:
-        with pytest.warns(
-                FutureWarning,
-                match="'stopping_strategy' attribute is deprecated"):
-            run([str(benchmark.benchmark_dir),
-                 *'-s solver1 -d test-dataset -n 1 -r 1 --no-plot'.split()],
-                standalone_mode=False)
-
-        run([str(benchmark.benchmark_dir),
-             *'-s solver2 -d test-dataset -n 1 -r 1 --no-plot'.split()],
-            standalone_mode=False)
-
-
-def test_deprecated_support_sparse():
-    # XXX remove in 1.5
-    assert benchopt.__version__ < '1.5'
-
-    solver1 = """from benchopt import BaseSolver
-    import numpy as np
-
-    class Solver(BaseSolver):
-        name = 'solver1'
-        support_sparse = True
-
-        def run(self, n_iter): pass
-
-        def set_objective(self, X, y, lmbd):
-            self.n_features = X.shape[1]
-
-        def get_result(self, **data):
-            return np.zeros(self.n_features)
-    """
-
-    with temp_benchmark(solvers=solver1) as benchmark:
-        with pytest.warns(
-                FutureWarning,
-                match="`support_sparse = False` is deprecated"):
-            run([str(benchmark.benchmark_dir),
-                 *'-s solver1 -d test-dataset -n 1 -r 1 --no-plot'.split()],
-                standalone_mode=False)
